Change Dijkstra's path finding to A*
Dijkstra's algorithm was painfully slow in this particular case. I need
to be able to quickly check if there are any paths for 4 pawns on a 9x9
boards for every possible (max. 128) fence placement, so that's 512 runs
of the algorithm in a row, and that's just to provide the player with a
list of all available valid moves. I noticed a over 50x faster
performance when switching to A*.